+++ build.gradle
@@ -11,18 +11,18 @@ buildscript {
         }
     }
     dependencies {
-        classpath 'net.minecraftforge.gradle:ForgeGradle:1.2-SNAPSHOT'
+        classpath 'net.minecraftforge.gradle:ForgeGradle:2.0-SNAPSHOT'
     }
 }
 
-apply plugin: 'forge'
+apply plugin: 'net.minecraftforge.gradle.forge'
 
-version = "1.8-0.0.7_12"
+version = "1.8-0.0.9_pre1"
 group= "com.comze.sanman00.expotions" // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = "expotions"
 
 minecraft {
-    version = "1.8-11.14.3.1496"
+    version = "1.8-11.14.4.1577"
     runDir = "eclipse"
     
     // the mappings can be changed at any time, and must be in the following format.
@@ -30,7 +30,7 @@ minecraft {
     // stable_#            stables are built at the discretion of the MCP team.
     // Use non-default mappings at your own risk. they may not allways work.
     // simply re-run your setup task after changing the mappings to update your workspace.
-    mappings = "snapshot_20150724"
+    mappings = "stable_18"
 }
 
 dependencies {
