+++ src/main/java/com/comze/sanman00/mods/minecraft/expotions/entity/EntityExplosivePotion.java
@@ -5,10 +5,7 @@ import net.minecraft.entity.projectile.EntityThrowable;
 import net.minecraft.util.BlockPos;
 import net.minecraft.util.MovingObjectPosition;
 import net.minecraft.world.World;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
 
-@SideOnly(Side.CLIENT)
 public class EntityExplosivePotion extends EntityThrowable {
 	public EntityExplosivePotion(World worldIn) {
 		super(worldIn);
@@ -27,7 +24,7 @@ public class EntityExplosivePotion extends EntityThrowable {
 		World world = super.worldObj;
 		
 		if (!world.isRemote && pos != null && pos.getBlockPos() != null) {
-			world.newExplosion(null, pos.getBlockPos().getX(), pos.getBlockPos().getY(), pos.getBlockPos().getZ(), 5.0f, false, true);
+			world.newExplosion(null, pos.getBlockPos().getX(), pos.getBlockPos().getY(), pos.getBlockPos().getZ(), 100.0f, false, true);
 			
             this.worldObj.playAuxSFX(2002, new BlockPos(this), 1);
             this.setDead();
