+++ src/main/java/com/comze/sanman00/mods/minecraft/expotions/proxy/ClientProxy.java
@@ -13,39 +13,24 @@ import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.fml.client.registry.RenderingRegistry;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-import net.minecraftforge.fml.common.registry.EntityRegistry;
 import net.minecraftforge.fml.common.registry.GameRegistry;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
+/**
+ * The proxy that loads things that cannot be loaded on the
+ * server. Things that cannot be loaded include modelsm etc,
+ */
 @SideOnly(Side.CLIENT)
-public class ClientProxy extends BaseProxy {
-    @Override
-    public void preInit(FMLPreInitializationEvent event) {
-        super.preInit(event);
-        Main.getLogger().info("Starting client-side pre-init of Explosive Potions mod version " + Main.VERSION);
-        GameRegistry.registerItem(ItemExplosivePotion.instance, "potion_explosive");
-        GameRegistry.registerItem(ItemThrowableExplosivePotion.instance, "potion_throwable_explosive");
-        EntityRegistry.registerModEntity(EntityExplosivePotion.class, "ThrowableExplosivePotion", 690, Main.instance, 2, 5, true);
-    }
-    
-    @Override
-    public void init(FMLInitializationEvent event) {
-        super.init(event);
+public class ClientProxy extends CommonProxy {
+	public void init(FMLInitializationEvent event) {
         Main.getLogger().info("Starting client-side init of Explosive Potions mod version " + Main.VERSION);
         Minecraft.getMinecraft().getRenderItem().getItemModelMesher().register(ItemExplosivePotion.instance, 0, new ModelResourceLocation("expotions:potion_explosive", "inventory"));
         Minecraft.getMinecraft().getRenderItem().getItemModelMesher().register(ItemThrowableExplosivePotion.instance, 0, new ModelResourceLocation("expotions:potion_throwable_explosive", "inventory"));
         GameRegistry.addShapelessRecipe(new ItemStack(ItemExplosivePotion.instance), new ItemStack(Items.glass_bottle), new ItemStack(Items.gunpowder));
         RenderingRegistry.registerEntityRenderingHandler(EntityExplosivePotion.class, new RenderExplosivePotion());
         MinecraftForge.EVENT_BUS.register(this);
-    }
-    
-    @Override
-    public void postInit(FMLPostInitializationEvent event) {
-        super.postInit(event);
     }
     
     //Fix for throwable potion not working
