--- src/main/java/com/comze/sanman00/mods/minecraft/expotions/proxy/ServerProxy.java
@@ -1,26 +1,0 @@
-package com.comze.sanman00.mods.minecraft.expotions.proxy;
-
-import net.minecraftforge.fml.common.event.FMLInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
-import net.minecraftforge.fml.relauncher.Side;
-import net.minecraftforge.fml.relauncher.SideOnly;
-
-@SideOnly(Side.SERVER)
-public class ServerProxy extends BaseProxy {
-    @Override
-    public void preInit(FMLPreInitializationEvent event) {
-        super.preInit(event);
-    }
-    
-    @Override
-    public void init(FMLInitializationEvent event) {
-        super.init(event);
-        
-    }
-    
-    @Override
-    public void postInit(FMLPostInitializationEvent event) {
-        super.postInit(event);
-    }
-}
